refactor(utils): 将getModelId函数从config/models迁移到utils模块
