+++ src/renderer/src/aiCore/middleware/feat/ImageGenerationMiddleware.ts
@@ -1,7 +1,8 @@
 import { BaseApiClient } from '@renderer/aiCore/clients/BaseApiClient'
-import { getModelId, isDedicatedImageGenerationModel } from '@renderer/config/models'
+import { isDedicatedImageGenerationModel } from '@renderer/config/models'
 import FileManager from '@renderer/services/FileManager'
 import { ChunkType } from '@renderer/types/chunk'
+import { getModelId } from '@renderer/utils'
 import { findImageBlocks, getMainTextContent } from '@renderer/utils/messageUtils/find'
 import { defaultTimeout } from '@shared/config/constant'
 import OpenAI from 'openai'
