+++ src/renderer/src/components/Popups/SelectModelPopup/popup.tsx
@@ -2,11 +2,12 @@ import { PushpinOutlined } from '@ant-design/icons'
 import ModelTagsWithLabel from '@renderer/components/ModelTagsWithLabel'
 import { TopView } from '@renderer/components/TopView'
 import { DynamicVirtualList, type DynamicVirtualListRef } from '@renderer/components/VirtualList'
-import { getModelId, getModelLogo, isEmbeddingModel, isRerankModel } from '@renderer/config/models'
+import { getModelLogo, isEmbeddingModel, isRerankModel } from '@renderer/config/models'
 import { usePinnedModels } from '@renderer/hooks/usePinnedModels'
 import { useProviders } from '@renderer/hooks/useProvider'
 import { getModelUniqId } from '@renderer/services/ModelService'
 import { Model, Provider } from '@renderer/types'
+import { getModelId } from '@renderer/utils'
 import { classNames, filterModelsByKeywords, getFancyProviderName } from '@renderer/utils'
 import { Avatar, Divider, Empty, Modal } from 'antd'
 import { first, sortBy } from 'lodash'
