+++ src/renderer/src/config/models.ts
@@ -152,7 +152,7 @@ import {
   ThinkingModelType,
   ThinkingOptionConfig
 } from '@renderer/types'
-import { getLowerBaseModelName, isUserSelectedModelType } from '@renderer/utils'
+import { getLowerBaseModelName, getModelId, isUserSelectedModelType } from '@renderer/utils'
 import OpenAI from 'openai'
 
 import { WEB_SEARCH_PROMPT_FOR_OPENROUTER } from './prompts'
@@ -3147,24 +3147,4 @@ export const isNotSupportSystemMessageModel = (model: Model): boolean => {
   }
 
   return false
-}
-
-/**
- * 获取模型ID，包含自定义ID的情况。用于模型类型判断、能力判断等场景，不要在网络通信API中使用。
- * @param model - 模型对象
- * @returns 返回模型的标准化ID字符串
- * @description
- * - 如果模型有自定义ID(customId)且不为空,则返回其小写的基础名称
- * - 否则返回模型原始ID的小写基础名称
- * - 使用 getLowerBaseModelName 进行标准化处理
- */
-export const getModelId = (model?: Model): string => {
-  if (!model) {
-    return ''
-  }
-  if (model.customId?.trim()) {
-    return getLowerBaseModelName(model.customId)
-  } else {
-    return getLowerBaseModelName(model.id)
-  }
 }
