+++ src/renderer/src/pages/settings/ProviderSettings/ModelList/ModelListItem.tsx
@@ -1,9 +1,10 @@
 import { type HealthResult, HealthStatusIndicator } from '@renderer/components/HealthStatusIndicator'
 import { HStack } from '@renderer/components/Layout'
 import ModelIdWithTags from '@renderer/components/ModelIdWithTags'
-import { getModelId, getModelLogo } from '@renderer/config/models'
+import { getModelLogo } from '@renderer/config/models'
 import { Model } from '@renderer/types'
 import { ModelWithStatus } from '@renderer/types/healthCheck'
+import { getModelId } from '@renderer/utils'
 import { maskApiKey } from '@renderer/utils/api'
 import { Avatar, Button, Tooltip } from 'antd'
 import { Bolt, Minus } from 'lucide-react'
